Refactored to share private keys
The bulk of this is refactoring to use a single private key for both
signing and encryption. Currently, we're just storing the Ed25519
private key, as we can use it to compute the X25519 key, but we probably
will want to store both in the future.

The other change is simplifying the notarization process after @dAxpeDDa
and I brainstormed a bit more. Ed25519 is already collision resistant,
so we're embracing that for now. The notarization will likely be
refactored to allow more than one signature to be notarized in one
payload, which allow for "stronger" notarizations.