Started working through the key server schema
In the encryption crate, the changes are mostly superficial -- to help
separate what is stored in the database to generate the unique id (the
primary key in the database), the notarization happens for an
"attestation" which contains all the data being written to the "ledger",
in our case the database. The notarization payload includes the
attestation, the id that the attestation is registered with the notary
as, and the notarization signature.

The data contained in the actual notarization payload remains unchanged.

The last update is allowing for multiple signatures to be submitted.
Right now it's completely pointless, since we only have a single
deterministic algorithm. But, we can add other algorithms, and by
signing with more than one algorith, the notarization and signature
become stronger.

That being said, they're already collision resilliant, so practical
attacks are not feasible even with just one signature.